async fetch lists w todos!
